refactor [client]: rename hooks folder and api hook
